+++ lib/volt/builder/output/file.ex
@@ -6,4 +6,12 @@ defmodule Volt.Builder.OutputFile do
             assets: [],
             chunk_id: nil,
             type: nil
+
+  @type t :: %__MODULE__{
+          path: String.t(),
+          size: non_neg_integer(),
+          assets: [term()],
+          chunk_id: String.t() | nil,
+          type: :entry | :chunk | nil
+        }
 end